requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
anytree==2.8.0
arxiv==1.4.7
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autocuda==0.16
beautifulsoup4==4.12.2
bert-score==0.3.13
blis==0.7.9
boostaug==2.3.5
boto3==1.26.156
botocore==1.29.156
bpemb==0.3.4
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.4
cloudpickle==2.2.1
confection==0.0.4
conllu==4.5.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.4.0
Deprecated==1.2.14
dill==0.3.5.1
docopt==0.6.2
editdistance==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.97.0
feedparser==6.0.10
ffmpy==0.3.0
filelock==3.12.2
findfile==2.0.0
flair==0.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gdown==4.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.35.2
gradio_client==0.2.7
grpcio==1.54.2
h11==0.14.0
h5py==3.9.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
hyperopt==0.2.7
idna==3.4
Janome==0.4.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keras==2.13.1rc0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
language-tool-python==2.7.1
lemminflect==0.2.3
libclang==16.0.0
linkify-it-py==2.0.2
lru-dict==1.2.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
metric-visualizer==0.9.7
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.13
murmurhash==1.0.9
natsort==8.4.0
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
num2words==0.5.12
numpy==1.25.0
oauthlib==3.2.2
OpenHowNet==2.0
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.1
packaging==23.1
pandas==2.0.2
pathy==0.10.2
Pillow==9.5.0
pinyin==0.4.0
pptree==3.1
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.5
py4j==0.10.9.7
pyabsa==2.3.1
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycld2==0.41
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-warmup==0.1.1
pytorch_revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
semantic-version==2.10.0
sentencepiece==0.1.99
seqeval==1.2.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlitedict==2.1.0
srsly==2.4.6
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-hub==0.13.0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
terminaltables==3.1.10
textattack @ git+https://github.com/yangheng95/TextAttack@0f6d47c13cc971c74d3f6a40882eb7d2512f7802
thinc==8.1.10
threadpoolctl==3.1.0
tikzplotlib==0.10.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.4
transformers==4.30.2
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
update-checker==0.18.0
urllib3==1.26.16
uvicorn==0.22.0
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
websockets==11.0.3
Werkzeug==2.3.6
Wikipedia-API==0.5.8
word2number==1.1
wrapt==1.15.0
XlsxWriter==3.1.2
xxhash==3.2.0
yarl==1.9.2
